// src/app/page/master/map-ruangan-jurusan/map-ruangan-jurusan.interface.ts
import {EventEmitter} from '@angular/core';

export interface mapRuanganJurusan {
	kdJurusan?: string;
	kdRuangan?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}